requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cmake==4.0.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
deepface==0.0.93
dlib==19.24.8
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
fastapi==0.110.1
fer==22.5.1
ffmpeg==1.4
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
keras-facenet==0.3.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.1
moviepy==1.0.3
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.9
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==10.2.0
proglog==0.1.11
protobuf==5.29.4
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.37.2
sympy==1.14.0
tbb==2021.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.7.0
torchvision==0.17.2
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.29.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
